Be more specific for project declared verions
In many cases firebase versions between modules are not the same. This will allow users to set a specific version for each module used. 

Example.
Firebase Core has a version 17.0.0 and a 17.0.1, but Firebase Messaging has 17.0.0, 17.0.1,17.1.0, 17.3.0, and more. It is possible to desire a version of firebase-messaging that doesn't have a mapping core version. 

This mainly comes up when trying to define project versions with react-native-push-notification since it looks are the same variable names, but for different modules.